refactor(structure): 구조 변경으로 인해 변경
LottoPurchaseInfo, LottoNumbersInfo로 분류됨으로 인해 변경했습니다.
